requirements.txt
python-dotenv
langchain
langchain_community
streamlit
langchain-mistralai
beautifulsoup4
pypdf
langchain-text-splitters
langchain-chroma
